// src/api-services/models/major-info-vo.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 学工大系统API
 * 学工大系统 Issues Documentation
 *
 * OpenAPI spec version: v1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 当前页数据记录，为查询类型的list集合
 * @export
 * @interface MajorInfoVo
 */
export interface MajorInfoVo {
    /**
     * id
     * @type {number}
     * @memberof MajorInfoVo
     */
    id?: number;
    /**
     * 学校名称
     * @type {string}
     * @memberof MajorInfoVo
     */
    schoolName?: string;
    /**
     * 学校链接
     * @type {string}
     * @memberof MajorInfoVo
     */
    schoolLink?: string;
    /**
     * 专业名称
     * @type {string}
     * @memberof MajorInfoVo
     */
    name?: string;
    /**
     * 专业排名
     * @type {string}
     * @memberof MajorInfoVo
     */
    ranking?: string;
    /**
     * 专业录取条件
     * @type {string}
     * @memberof MajorInfoVo
     */
    admissionRequirement?: string;
    /**
     * 专业申请时间
     * @type {Date}
     * @memberof MajorInfoVo
     */
    applicationTime?: Date;
    /**
     * 专业课程详情
     * @type {string}
     * @memberof MajorInfoVo
     */
    courseDetails?: string;
    /**
     * 专业链接
     * @type {string}
     * @memberof MajorInfoVo
     */
    link?: string;
    /**
     * 专业申请通道
     * @type {string}
     * @memberof MajorInfoVo
     */
    applicationChannel?: string;
}