// src/api-services/models/page-major-info-vo.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 学工大系统API
 * 学工大系统 Issues Documentation
 *
 * OpenAPI spec version: v1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { MajorInfoVo } from './major-info-vo';
/**
 * 
 * @export
 * @interface PageMajorInfoVo
 */
export interface PageMajorInfoVo {
    /**
     * 当前页
     * @type {number}
     * @memberof PageMajorInfoVo
     */
    current?: number;
    /**
     * 首页
     * @type {number}
     * @memberof PageMajorInfoVo
     */
    first?: number;
    /**
     * 上一页
     * @type {number}
     * @memberof PageMajorInfoVo
     */
    pre?: number;
    /**
     * 下一页
     * @type {number}
     * @memberof PageMajorInfoVo
     */
    next?: number;
    /**
     * 尾页
     * @type {number}
     * @memberof PageMajorInfoVo
     */
    last?: number;
    /**
     * 每页最大记录数
     * @type {number}
     * @memberof PageMajorInfoVo
     */
    pageSize?: number;
    /**
     * 总记录数
     * @type {number}
     * @memberof PageMajorInfoVo
     */
    total?: number;
    /**
     * 总页数
     * @type {number}
     * @memberof PageMajorInfoVo
     */
    pages?: number;
    /**
     * 当前页实际记录数
     * @type {number}
     * @memberof PageMajorInfoVo
     */
    currSize?: number;
    /**
     * 当前页数据记录，为查询类型的list集合
     * @type {Array<MajorInfoVo>}
     * @memberof PageMajorInfoVo
     */
    list?: Array<MajorInfoVo>;
}