// src/api-services/models/page-school-info.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 学工大系统API
 * 学工大系统 Issues Documentation
 *
 * OpenAPI spec version: v1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { SchoolInfo } from './school-info';
/**
 * 
 * @export
 * @interface PageSchoolInfo
 */
export interface PageSchoolInfo {
    /**
     * 当前页
     * @type {number}
     * @memberof PageSchoolInfo
     */
    current?: number;
    /**
     * 首页
     * @type {number}
     * @memberof PageSchoolInfo
     */
    first?: number;
    /**
     * 上一页
     * @type {number}
     * @memberof PageSchoolInfo
     */
    pre?: number;
    /**
     * 下一页
     * @type {number}
     * @memberof PageSchoolInfo
     */
    next?: number;
    /**
     * 尾页
     * @type {number}
     * @memberof PageSchoolInfo
     */
    last?: number;
    /**
     * 每页最大记录数
     * @type {number}
     * @memberof PageSchoolInfo
     */
    pageSize?: number;
    /**
     * 总记录数
     * @type {number}
     * @memberof PageSchoolInfo
     */
    total?: number;
    /**
     * 总页数
     * @type {number}
     * @memberof PageSchoolInfo
     */
    pages?: number;
    /**
     * 当前页实际记录数
     * @type {number}
     * @memberof PageSchoolInfo
     */
    currSize?: number;
    /**
     * 当前页数据记录，为查询类型的list集合
     * @type {Array<SchoolInfo>}
     * @memberof PageSchoolInfo
     */
    list?: Array<SchoolInfo>;
}