// src/api-services/models/school-past-ranking-vo.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 学工大系统API
 * 学工大系统 Issues Documentation
 *
 * OpenAPI spec version: v1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface SchoolPastRankingVo
 */
export interface SchoolPastRankingVo {
    /**
     * 
     * @type {number}
     * @memberof SchoolPastRankingVo
     */
    schoolId?: number;
    /**
     * 
     * @type {string}
     * @memberof SchoolPastRankingVo
     */
    schoolName?: string;
    /**
     * 
     * @type {number}
     * @memberof SchoolPastRankingVo
     */
    year?: number;
    /**
     * 
     * @type {number}
     * @memberof SchoolPastRankingVo
     */
    rank?: number;
}